機能追加: ProjectViewContextService の実装（プロジェクト管理機能） プロジェクト関連の権限・表示情報を管理する ProjectViewContextService を新規作成
ユーザーIDでプロジェクト情報を取得する処理に MyBatis を統合

プロジェクト詳細画面における Model 属性の整理と権限情報の出力を追加

ProjectRepositoryService から MyBatis 依存を除去し、JPAベースにリファクタリング

DBInfoMapper および UserMapper を DTO / JPA エンティティベースに変更

UserEntity および UserJpaRepository を新規作成し、ユーザー管理機能を刷新

UserController および UserService を新しいユーザー構成に対応

Thymeleaf テンプレート（プロジェクト詳細・ナビゲーション）を権限連携に対応

ラベルやプロパティファイルを追加

+++ src/main/java/com/example/projectmanagement/modules/coding/datastructure/entity/EntityEntity.java
@@ -10,12 +10,12 @@ import lombok.NoArgsConstructor;
 public class EntityEntity {
 
 	private Integer id;
-	private Integer tableColumnId;
-	private Integer fieldId;
+	private Long tableColumnId;
+	private Long fieldId;
 	private Integer projectId;
 	
 	
-	public EntityEntity(Integer tableColumnId, Integer fieldId, Integer projectId) {
+	public EntityEntity(Long tableColumnId, Long fieldId, Integer projectId) {
 		super();
 		this.tableColumnId = tableColumnId;
 		this.fieldId = fieldId;
